identify duplicates in data
